% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abs_abund}
\alias{abs_abund}
\title{An example absolute abundance data.frame}
\format{
A data.rame with 44 observations and 1 variable
}
\usage{
abs_abund
}
\description{
A data.frame where the first column is a numeric vector of absolute abundance values
for each sample in seqmat (rownames are identical to seqmat)
}
\keyword{datasets}
