% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{asv_clustr}
\alias{asv_clustr}
\title{Hierarchical clustering of ASVs.}
\usage{
asv_clustr(asv_list, agg_method = "ward.D2", td = "none", k = NULL)
}
\arguments{
\item{asv_list}{An asv_list.}

\item{agg_method}{A string of either "ward", "ward.D2", "complete", "single", "average", "mcquitty", "median", or "centroid".
"ward.D2" is the default. See ?hclust for more details.}

\item{td}{A string of either "none", "both", or "td_only". Default is "none":
"none" will use only abundance data to perform cosine similarity calculation.
"both" will use both the abundance data and time rate derivative data.
"td_only" will use only the time rate derivative data.}

\item{k}{Optional. An integer. The number of clusters you wish to group your ASVs into. If a k is not chosen asv_clustr will
return a dendrogram plot instead}
}
\value{
An asv_list object, or a dendrogram if no k is set
}
\description{
Calculates the cosine similarity, converts to an angular distance matrix, and then either
returns a dendrogram if no k is set, or returns an asv_list with an h_clust element.
}
\examples{
clustr_dendro  <- asv_clustr(asv_list = asv_list)
plot(clustr_dendro, labels = FALSE)
rect.hclust(tree = clustr_dendro, border ="red", k=4)

asv_list <- asv_clustr(asv_list = asv_list,  k=4)

asv_list <- asv_clustr(asv_list = asv_list,  k=4, td = "both")

asv_list <- asv_clustr(asv_list = asv_list,  k=4, td = "td_only")


}
