% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asv_list}
\alias{asv_list}
\title{An example asv_list}
\format{
A list with 5 elements.
}
\usage{
asv_list
}
\description{
A list containing the following : seqmat (a numeric matrix), taxmat (a character matrix),
meta (a data.frame), seqmat_td (a numeric matrix), and h_clust (a data.frame).
}
\keyword{datasets}
