% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cluster_means_bar}
\alias{cluster_means_bar}
\title{Bar plot of cluster means.}
\usage{
cluster_means_bar(melted_asv_list, time_var)
}
\arguments{
\item{melted_asv_list}{A data.frame. The output of melt_asv_list.}

\item{time_var}{A string. The metadata variable for timepoint.}
}
\value{
A ggplot2 object
}
\description{
Creates a barplot of cluster means. Useful for checking if cluster patterns are unique and if dispersion is low. Can add ggplot2 aesthetic layers.
}
\examples{
cluster_means_bar(melted_asv_list, time_var = "Timepoint")

}
