% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cluster_phytree}
\alias{cluster_phytree}
\title{Plot a cluster phytree.}
\usage{
cluster_phytree(asv_list, tree, cladogram = FALSE, layout = "circular")
}
\arguments{
\item{asv_list}{An asv_list. Must have an h_clust element.}

\item{tree}{A tree of phylo class produced with the seqmat given to the asv_list constructor function.}

\item{cladogram}{A boolean. TRUE produces a cladogram where branch lengths are equal, FALSE produces a
phylogenetic tree. Default is FALSE.}

\item{layout}{A string. See ?ggtree::ggtree for layout options. Default is "circular".}
}
\value{
A ggtree object
}
\description{
Wrapper function for ggtree that produces a tree with branches and tips colored by cluster. can use ggtree aesthetic layers.
}
\examples{
cluster_phytree(asv_list = asv_list, tree = phangtree$tree)

cluster_phytree(asv_list = asv_list, tree = phangtree$tree, cladogram = TRUE, layout = "radial")

}
