% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{compare_pd}
\alias{compare_pd}
\title{Compare population dynamics between groups.}
\usage{
compare_pd(
  asv_list,
  sam_var,
  time_var,
  independent_var,
  batch,
  by_asv = FALSE,
  rescale = FALSE
)
}
\arguments{
\item{asv_list}{An asv_list.}

\item{sam_var}{A string. The metadata variable for samples.}

\item{time_var}{A string. The metadata variable for timepoints.}

\item{independent_var}{A string. The metadata variable for the experimental variable(s).}

\item{batch}{Optional. A string. A metadata variable for a batch effect. The batch effect must be separate
from the independent variable, i.e. levels of the batch variable must be distributed among the levels of
the idependent variable.}

\item{by_asv}{A boolean. If TRUE, difference in integral means will be computed for every ASV. If FALSE
difference in integral means will be computed for each cluster. Default is FALSE}

\item{rescale}{A boolean. If TRUE, each ASV for each sample will be rescaled to a maximum value of 1 and
minimum value of 0. Otherwise raw read counts will be used. Default is FALSE}
}
\value{
An asv_list.
}
\description{
Compares mean population dynamics of either clusters to determine if an experimental variable
or batch introduces noise, or individual ASVs to determine if an independent variable effects population dynamics.
}
\examples{

\dontshow{
new_meta <- asv_list$meta

new_meta$batch <- c("A","B","C","D")

asv_list$meta <- new_meta

}
pd_cage_effect <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint",independent_var = "Treatment", batch="batch")

pd_cage_effect_rescale <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment", batch="batch", rescale = TRUE)

pd_cage_effect_ASV <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment", batch="batch", by_asv = TRUE)

pd_cage_effect_ASV_rescale <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint",independent_var = "Treatment",
batch="batch", by_asv = TRUE, rescale = TRUE)

pd_Treatment <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment")

pd_Treatment_rescale <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment", rescale = TRUE)

pd_Treatment_ASV <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment", by_asv = TRUE)

pd_Treatment_ASV_rescale <- compare_pd(asv_list = asv_list, sam_var = "Sample",
time_var = "Timepoint", independent_var = "Treatment", by_asv = TRUE, rescale = TRUE)


}
