% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{make_asv_list}
\alias{make_asv_list}
\title{Creation of an asv_list list object.}
\usage{
make_asv_list(seqmat, taxmat, meta)
}
\arguments{
\item{seqmat}{A numeric matrix of sequencing reads, with ASVs as columns and unique sample identifiers as rownames.
The correct format is that of a dada2 sequence count matrix.}

\item{taxmat}{A character matrix of taxonomic information, with rownames matching seqmat column names, and taxonomic
rankings (Kingdom, Phylum, Class etc) as columns. The correct format is that of a dada2 taxonomy matrix.}

\item{meta}{A data frame with metadata for your sequencing samples. Rownames should be equal and in identical order to the seqmat.
It also msut contain the following: a column of non unique sample identifiers (i.e. mouse number, plant number, etc),
a column of time points as numeric, a column of independent variable(s) or grouped independent variables is optional.}
}
\value{
A list with seqmat, taxmat, meta. additionally adds the class ASVclustr
}
\description{
Constructor function that creates a list with the added class ASVclustr. The list needs to contain at least a sequence matrix, taxa matrix, and
metadata data.frame.
}
\examples{
asv_list_strep <- make_asv_list(seqmat = seqmat, taxmat = taxmat, meta = meta)

}
