% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{melt_asv_list}
\alias{melt_asv_list}
\title{Melt an asv_list into a long form data.frame for plotting functions.}
\usage{
melt_asv_list(asv_list, ratio = FALSE, rescale = FALSE, sam_var, time_var)
}
\arguments{
\item{asv_list}{An asv_list.}

\item{ratio}{A boolean. TRUE will convert reads into a ratio for each sample. Default is FALSE.}

\item{rescale}{A boolean. TRUE will rescale reads for each ASV in each sample to a value between 0 (min) and 1 (max).}

\item{sam_var}{A string. The name of the sample variable in metadata.}

\item{time_var}{A string. The name of the timepoint variable in metadata.}
}
\value{
A data.frame in longform with every variable as a single vector for use in plotting functions.
}
\description{
Creates a data.frame from an asv list in long-form with each variable as a single vector.
}
\examples{
melted_asv <- melt_asv_list(asv_list = asv_list, sam_var = "Sample", time_var = "Timepoint")

}
