% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melted_asv_list}
\alias{melted_asv_list}
\title{An example melted_asv_list data.frame}
\format{
3168 observations of 17 variables
}
\usage{
melted_asv_list
}
\description{
A dataframe in longform. Suitable for use with ggplot2.
}
\keyword{datasets}
