% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{An example metadata data.frame}
\format{
A data.frame with 44 observations and 7 variables
}
\usage{
meta
}
\description{
A data.frame with metadata for the samples (rows) in seqmat
}
\keyword{datasets}
