% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{order_seqmat}
\alias{order_seqmat}
\title{Order seqmat by sample and timepoint}
\usage{
order_seqmat(seqmat, meta, sam_var, time_var)
}
\arguments{
\item{seqmat}{seqmat from an asv_list}

\item{meta}{metadata from an asv_list}

\item{sam_var}{sample variable from metadata}

\item{time_var}{timepoint variable from metadata}
}
\value{
a list of seqmats for every sample ordered by timepoint
}
\description{
Internal function to order a seqmat object by time point and convert to a list of dataframes with each sample as an element.
}
\examples{
seqmat <- as.data.frame(asv_list$seqmat)
meta <- asv_list$meta
seqmat_order_list <- order_seqmat(seqmat = seqmat, meta = meta, sam_var = "Sample",
time_var = "Timepoint")


}
