% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phangtree}
\alias{phangtree}
\title{An example phylo tree object}
\format{
A pml object with 22 elements.
}
\usage{
phangtree
}
\description{
A pml object with a phylo element named 'tree' that was generated from the seqmat example data by
making a 16S gene alignment using the package DECIPHER and making a phylogenetic tree from the
alignment using the package phangorn.
}
\keyword{datasets}
