% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples}
\usage{
remove_samples(asv_list, variable, variable_level)
}
\arguments{
\item{asv_list}{An asv_list object.}

\item{variable}{Optional. A string of a variable with levels you want to remove.}

\item{variable_level}{A string or character vector. If the variable parameter is used, a string of the variable level you wish to remove.
Otherwise, a character vector of sample names you wish to remove.}
}
\value{
An asv_list object.
}
\description{
Either removes samples by a variable level, or removes samples by name.
}
\examples{

asv_list_pruned <- remove_samples(asv_list=asv_list,
variable_level = c("EVelazquez0717_480.1.B", "EVelazquez0717_480.1.C"))

asv_list_dirty <- remove_samples(asv_list = asv_list, variable = "Treatment",
variable_level = "Dirty-A")

}
