% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{remove_taxa}
\alias{remove_taxa}
\title{Remove taxa}
\usage{
remove_taxa(asv_list, taxa)
}
\arguments{
\item{asv_list}{An asv_list object.}

\item{taxa}{A string or character vector. The taxa to remove, e.g. "Clostridia", "Escherichia", etc.}
}
\value{
an asv_list object
}
\description{
Remove ASVs from seqmat by taxonomic value.
}
\examples{

asv_list_subset_abund <- remove_samples(asv_list=asv_list,
variable_level = c("EVelazquez0717_480.1.B", "EVelazquez0717_480.1.C"))

asv_list_subset_<- remove_samples(asv_list = asv_list, variable = "Treatment",
variable_level = "Dirty-A")


}
