% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqmat}
\alias{seqmat}
\title{An example sequence count matrix}
\format{
A numeric matrix with 44 rows and 323 columns
}
\usage{
seqmat
}
\description{
A numeric matrix of sequence read counts.
}
\keyword{datasets}
