% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{subset_ASVs}
\alias{subset_ASVs}
\title{Subset ASVs by read count and presence.}
\usage{
subset_ASVs(asv_list, sam_var, sam_threshold, abund_threshold)
}
\arguments{
\item{asv_list}{An asv_list object.}

\item{sam_var}{A string. The name of the sample variable in metadata.}

\item{sam_threshold}{An integer. Threshold for the number of samples in which an ASV is present.}

\item{abund_threshold}{Optional. An integer. Threshold for the amount of reads present for an ASV in the dataset.}
}
\value{
an asv_list object
}
\description{
Subsets to ASVs that reach a required threshold of reads and or presence across samples.
}
\examples{

asv_list_subset_abund <- remove_samples(asv_list=asv_list,
variable_level = c("EVelazquez0717_480.1.B", "EVelazquez0717_480.1.C"))

asv_list_subset_<- remove_samples(asv_list = asv_list, variable = "Treatment",
variable_level = "Dirty-A")


}
