% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{sum_by_cluster}
\alias{sum_by_cluster}
\title{Sum by cluster}
\usage{
sum_by_cluster(asv_list, abs_abund)
}
\arguments{
\item{asv_list}{An asv list with an h_clust element.}

\item{abs_abund}{Optional. A data.frame with matching rownames to seqmat and meta within the asv_list. The first
column must be a numeric vector of total abundance values for each unique sample.}
}
\value{
A data.frame in longform with normalized cluster abundances for each sample.
}
\description{
Sums ASVs by cluster.
}
\examples{

clustersum <- sum_by_cluster(asv_list = asv_list)

clustersum_total <- sum_by_cluster(asv_list = asv_list, abs_abund = abs_abund)

}
