% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{taxa_by_cluster}
\alias{taxa_by_cluster}
\title{Investigate taxa abundance by cluster.}
\usage{
taxa_by_cluster(
  melted_asv_list,
  tax_level,
  time_var,
  sam_var,
  plot_boxplot = FALSE
)
}
\arguments{
\item{melted_asv_list}{A data.frame. Produced from the function melt_asv_list.}

\item{tax_level}{A string. The taxonomic rank you want to aggregate by.}

\item{time_var}{A string. The timepoint variable in meta.}

\item{sam_var}{A string. The sample variable in meta.}

\item{plot_boxplot}{Optional. A boolean, defualt is FALSE. If FALSE,
returns a data.frame. Otherwise returns a ggplot2 boxplot}
}
\value{
A data.frame if plot_boxplot = FALSE, or a boxplot if plot_boxplot = TRUE
}
\description{
Aggregates taxa by sample, cluster, and timepoint by a specified taxonomic level.
}
\examples{

taxa_by_cluster(melted_asv_list = melted_asv_list, tax_level = Class, time_var = Timepoint,
sam_var = Sample)

taxa_by_cluster(melted_asv_list = melted_asv_list, tax_level = Class, time_var = Timepoint,
sam_var = Sample, plot_boxplot = TRUE)


}
