% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxmat}
\alias{taxmat}
\title{An example taxonomy matrix}
\format{
A character matrix with 378 rows and 6 columns
}
\usage{
taxmat
}
\description{
A character matrix with taxonomy data for the ASVs (columns) in seqmat
}
\keyword{datasets}
