% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{calc_td}
\alias{calc_td}
\title{Calculation of time rate derivatives.}
\usage{
calc_td(asv_list, sam_var, time_var)
}
\arguments{
\item{asv_list}{An asv_list.}

\item{sam_var}{A string. The metadata variable for samples.}

\item{time_var}{A string. The metadata variable for timepoints.}
}
\value{
An asv_list.
}
\description{
Computes the time rate derivatives for every ASV in every sample and adds them as an element in your asv_list object asv_list.
}
