% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{compare_pd}
\alias{compare_pd}
\title{compare population dynamics across samples}
\usage{
compare_pd(
  asv_list,
  batch = FALSE,
  calc_CI = FALSE,
  boot_k = NULL,
  groups = NULL
)
}
\arguments{
\item{asv_list}{an asv_list object}

\item{batch}{a boolean. TRUE if stratifying out a batch effect from another covariate, otherwise FALSE}

\item{calc_CI}{a boolean. TRUE will calculate 95 percent confidence intervals using non parametric bootstrapping}

\item{boot_k}{an integer. default is NULL. If calc_CI is TRUE, boot_k is the number of bootstrap replicates
to perform. if sample sizes are insufficient for number of bootstraps, a vector of ASVs or clusters that had
insufficient sample size will be returned}

\item{groups}{a character vector. default is NULL. a character vector of strings denoting levels of an
independent covariate. the first string in the vector will be used as the reference group that all
other groups will be compared to.}
}
\value{
a data.frame
}
\description{
Compare cluster orASV integral means using a binomial-gamma hurdle model on median of ratios normalized
sequencing counts
}
